bot/requirements.txt
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
discord.py==1.6.0
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.52.0
grpcio==1.35.0
hiredis==1.1.0
idna==2.10
libcst==0.3.16
loguru==0.5.3
multidict==5.1.0
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
urllib3==1.26.3
win32-setctime==1.0.3
yarl==1.6.3
Pillow==8.2.0
python-dotenv~=0.17.1